+++ 2. Modulo Spring/ej_concesionaria/src/main/java/com/example/concesionaria/services/VehiculosService.java
@@ -15,6 +15,6 @@ public interface VehiculosService {
 
     List<Vehiculo> getByPrice(Integer minimo, Integer maximo);
 
-    Vehiculo getById(Integer id);
+    Vehiculo getById(Long id);
 
 }
